Separate three-agent view verification concerns
The green disclosure implementation pushed the shared three-agent verifier over the repo size limit. Moving view-specific privacy and disclosure checks into a focused module keeps commitment verification small while preserving the same fail-closed behavior.

Constraint: AGENTS.md requires files under 200 lines and a mandatory refactor phase.

Rejected: Leave the verifier at 215 lines | that would violate the local code standard after #7051.

Confidence: high

Scope-risk: narrow

Directive: Keep disclosure checks in three_agent_view_verify.rs unless they become shared verifier primitives.

Tested: cargo fmt

Tested: wc -l crates/kamn-e2e-harness/src/mvp_demo/three_agent_claim.rs crates/kamn-e2e-harness/src/mvp_demo/three_agent_verify.rs crates/kamn-e2e-harness/src/mvp_demo/three_agent_view_verify.rs crates/kamn-e2e-harness/src/mvp_demo/report_markdown.rs crates/kamn-e2e-harness/tests/mvp_demo_three_agent_claim_contract.rs

Tested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo test -p kamn-e2e-harness --test mvp_demo_three_agent_claim_contract -- --nocapture

Not-tested: Full workspace gates and demo command not run yet.
