+++ crates/kamn-e2e-harness/src/mvp_demo/mod.rs
@@ -19,6 +19,7 @@ mod runner;
 mod service_api_proof;
 mod three_agent_claim;
 mod three_agent_verify;
+mod three_agent_view_verify;
 mod verify;
 mod verify_support;
